A dockerized build process for the portfolios site
It may not be ideal, but it eliminates the "works on my computer"
problem.

Instructions are in DEVELOPMENT.md

In short:

1. Build the container: docker build -t portfolios/build .
2. Build the site: docker run -ti --rm -v ${PWD}:/site portfolios/build:latest -b
3. Serve the site: docker run -ti --rm -p 4000:4000 -v ${PWD}:/site portfolios/build:latest -s

Those instructions work on Macs; the env variable may change for
Windows.

The important thing is that the build process happens inside the
container, but the local directory is mounted as the /site workspace.
This means that no files live in the container (including the entry.bash
script).